feat: improving loading time by moving wasm load to after first paint
