+++ website/src/components/WasmLoadingMessage.jsx
@@ -1,0 +1,28 @@
+import { App } from "antd";
+import { useAleoWASM } from "../aleo-wasm-hook.js";
+import { useEffect } from "react";
+
+export function WasmLoadingMessage() {
+    const { message, notification } = App.useApp();
+    const [_, aleoLoading] = useAleoWASM();
+
+    useEffect(() => {
+        if(aleoLoading){
+            message.open({
+                key: "wasmLoading",
+                type: 'loading',
+                content: 'Loading WASM...some functionality may not be available yet...',
+                duration: 0,
+            });
+        } else {
+            message.open({
+                key: "wasmLoading",
+                type: 'success',
+                content: 'WASM Loaded!',
+                duration: 2,
+            });
+        }
+    }, [aleoLoading]);
+
+    return null;
+}
